Add support for layout schematics, GIF/Video wallpapers, and invisible buttons
- Implement LauncherSchematic model and repository for backup/restore.
- Fix color serialization by using Long in data models to prevent overflow.
- Add support for GIF and Video wallpapers using Coil for frame decoding.
- Implement InvisibleButton launcher item for transparent grid actions.
- Fix bug where secret wallpaper would bleed onto the lockscreen.
- Update UI pickers and menus to expose new wallpaper types and item.

+++ app/src/main/kotlin/com/quimodotcom/lqlauncher/compose/launcher/LiquidGlassSettingsScreen.kt
@@ -192,9 +192,9 @@ fun LiquidGlassSettingsScreen(
                     item {
                         ColorPickerSetting(
                             title = "Panel Tint Color",
-                            currentColor = Color(settings.panelTintColor),
+                            currentColor = Color(settings.panelTintColor.toInt()),
                             onColorSelected = { 
-                                onSettingsChanged(settings.copy(panelTintColor = it.toArgb()))
+                                onSettingsChanged(settings.copy(panelTintColor = it.toArgb().toLong()))
                             }
                         )
                     }
@@ -779,9 +779,9 @@ fun LiquidGlassSettingsScreen(
                     item {
                         ColorPickerSetting(
                             title = "Dot Color",
-                            currentColor = Color(settings.notificationDotColor),
+                            currentColor = Color(settings.notificationDotColor.toInt()),
                             onColorSelected = {
-                                onSettingsChanged(settings.copy(notificationDotColor = it.toArgb()))
+                                onSettingsChanged(settings.copy(notificationDotColor = it.toArgb().toLong()))
                             },
                             enabled = !settings.liquidGlassNotificationDots
                         )
